# Backend/notification/Swagger.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8086","description":"Generated server url"}],"paths":{"/api/notifications/{id}/read":{"put":{"tags":["notification-controller"],"operationId":"markAsRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseNotificationResponse"}}}}}}},"/api/notifications":{"get":{"tags":["notification-controller"],"operationId":"getNotifications","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListNotificationResponse"}}}}}},"post":{"tags":["notification-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseNotificationResponse"}}}}}}},"/api/notifications/unread-count":{"get":{"tags":["notification-controller"],"operationId":"unreadCount","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringLong"}}}}}}}},"components":{"schemas":{"ApiResponseNotificationResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/NotificationResponse"},"timestamp":{"type":"string","format":"date-time"}}},"NotificationResponse":{"type":"object","properties":{"notificationId":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"message":{"type":"string"},"category":{"type":"string"},"status":{"type":"string"},"createdDate":{"type":"string","format":"date-time"}}},"NotificationRequest":{"type":"object","properties":{"userId":{"type":"integer","format":"int32"},"message":{"type":"string","minLength":1},"category":{"type":"string"}},"required":["message","userId"]},"ApiResponseListNotificationResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseMapStringLong":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}},"timestamp":{"type":"string","format":"date-time"}}}}}}