# Backend/notification/src/main/java/com/cts/adstudio/notificationservice/dto/response/NotificationResponse.java
package com.cts.adstudio.notificationservice.dto.response;

import lombok.Builder;
import lombok.Data;
import java.time.LocalDateTime;

@Data
@Builder
public class NotificationResponse {
    private Integer notificationId;
    private Integer userId;
    private String message;
    private String category;
    private String status;
    private LocalDateTime createdDate;
}